+++ internal/tester/scenarios/s1enb/network_rule.go
@@ -124,7 +124,7 @@ func runS1ENBNetworkRule(ctx context.Context, env scenarios.Env, params *probePa
 
 	ue := e.NewUE(imsi, k, opc)
 	if ipv6 {
-		ue.RequestPDNType(eps.PDNTypeIPv6)
+		ue.RequestPDNType(uint8(eps.PDNTypeIPv6))
 	}
 
 	res, err := e.Attach(ue, 15*time.Second)
